+++ Batallon/src/modelo/rangos/Nivel_militar.java
@@ -1,0 +1,8 @@
+package modelo.rangos;  // Paquete de la clase  
+// Enumeración que define los niveles militares
+public enum Nivel_militar {
+    CAPITAN ,       // Nivel de Capitán 
+    SOLDADO_RASO,   // Nivel de Soldado Raso
+    TENIENTE,       // Nivel de Teniente
+    CORONEL         // Nivel de Coronel
+}
